requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
groq>=0.9.0
gspread==6.1.2
google-auth==2.29.0
pydantic==2.7.1
httpx>=0.27.0